progressing w/ cleaning up duckdb
